requirements.txt
pandas==2.0.3
numpy==1.24.3
scikit-learn==1.3.0
matplotlib==3.7.2
seaborn==0.12.2
plotly==5.16.1
jupyter==1.0.0
scipy==1.11.2
statsmodels==0.14.0
pyyaml==6.0
python-dotenv==1.0.0
notebook==6.5.4
ipython==8.14.0